requirements.txt
pyyaml>=6.0
mitmproxy>=8.0